Refactor to follow the standard Go package guide
- Module name is the repo name
- main package should be in cmd/<binary-name>/main.go
