+++ go.mod
@@ -1,3 +1,3 @@
-module paperswithcode
+module github.com/codingpot/paperswithcode-go
 
 go 1.16
